requirements.txt
py-cord==2.0.0b3
PyNaCl
colorlog
python-dotenv
jishaku